refactor navigation and state management using PageContext
