+++ src/pages/OnboardingPage/OnboardingPage1.jsx
@@ -1,23 +1,25 @@
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 import "./OnboardingPage.scss";
 import Button from "../../components/Button/Button";
+import { PageContext } from "../../contexts/PageContext";
 import loadingInactiveIcon from "../../assets/images/loadingInactiveIcon.svg";
 import loadingActiveIcon from "../../assets/images/loadingActiveIcon.svg";
 import arrowLeft from "../../assets/images/arrowLeft.svg";
 
-const OnboardingPage1 = ({ onNext, onBack, progressBarIndex }) => {
+const OnboardingPage1 = () => {
   const [selectedButton, setSelectedButton] = useState(null);
+  const { handleNext, handleBack, progressBarIndex } = useContext(PageContext);
 
   const handleClick = (button) => {
     setSelectedButton(button);
-    onNext();
+    handleNext();
   };
 
   return (
     <div className="onboarding-page">
       <div className="onboarding-page__container">
         <>
-          <div className="left-arrow-container" onClick={onBack}>
+          <div className="left-arrow-container" onClick={handleBack}>
             <img src={arrowLeft} alt="arrowLeft" className="arrow-left" />
           </div>
           <div className="onboarding-page__text-container">
